% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm3pc_sys.R
\name{abm3pc_sys}
\alias{abm3pc_sys}
\title{Adams-Bashford-Moulton for system of equations}
\usage{
abm3pc_sys(f, a, b, y0, n, Pars)
}
\arguments{
\item{f}{function in the differential equation y' = f(x, y);
defined as a function R \times R^m \rightarrow R^m, where m is the number of equations.}

\item{a}{starting time for the interval to integrate}

\item{b}{ending time for the interval to integrate.}

\item{y0}{starting values at time \code{a}}

\item{Pars}{named list of parameters passed to f}
}
\description{
Third-order Adams-Bashford-Moulton predictor-corrector method.
}
\details{
Combined Adams-Bashford and Adams-Moulton (or: multi-step) method of third order
with corrections according to the predictor-corrector approach.
Copied from pracma under GPL-3, with small modifications to work with RTMB
}
