% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_equilibrium.R
\name{add_equilibrium}
\alias{add_equilibrium}
\title{Compute equilibrium values}
\usage{
add_equilibrium(ecoparams, scale_solver, noB_i, type_i)
}
\arguments{
\item{ecoparams}{list of parameters}
}
\description{
Compute equilibrium values
}
\details{
todo
}
