% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanzas.R
\name{add_stanza_params}
\alias{add_stanza_params}
\title{Compute state-varaibles for multi-stanza groups}
\usage{
add_stanza_params(p, stanza_data, settings)
}
\description{
Taking parameters, settings, and output from `make_stanza_dat,
initialize the equilibrium (SplitAlpha) and initial state-variables
for multi-stanza groups
}
