% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nll.R
\name{compute_nll}
\alias{compute_nll}
\title{Compute negative log-likelihood for EcoState model}
\usage{
compute_nll(p)
}
\arguments{
\item{p}{list of parameters}
}
\description{
Compute negative log-likelihood for EcoState model
}
\details{
todo
}
