% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostate.R
\name{ecostate}
\alias{ecostate}
\title{fit EcoState model}
\usage{
ecostate(
  taxa,
  years,
  catch,
  biomass,
  PB,
  QB,
  B,
  DC,
  EE,
  V,
  type,
  U,
  fit_B = vector(),
  fit_Q = vector(),
  fit_B0 = vector(),
  fit_EE = vector(),
  fit_eps = vector(),
  control = ecostate_control()
)
}
\arguments{
\item{taxa}{Character vector of taxa included in model.}

\item{years}{Integer-vector of years included in model}

\item{catch}{long-form data frame with columns \code{Mass}, \code{Year}
and  \code{Taxon}}

\item{biomass}{long-form data frame with columns \code{Mass}, \code{Year}
and  \code{Taxon}, where \code{Mass} is assumed to have the same units
as \code{catch}}

\item{PB}{numeric-vector with names matching \code{taxa}, providing the
ratio of production to biomass for each taxon}

\item{QB}{numeric-vector with names matching \code{taxa}, providing the
ratio of consumption to biomass for each taxon}

\item{B}{numeric-vector with names matching \code{taxa}, providing the
starting (or fixed) value for equilibrium biomass for each taxon}

\item{DC}{numeric-matrix with rownames and colnamesmatching \code{taxa},
where each column provides the diet proportion for a given predator}

\item{V}{numeric-matrix with rownames and colnamesmatching \code{taxa},
where each element gives the vulnerability parameter for a given
interaction.}

\item{type}{character-vector with names matching \code{taxa} and
elements \code{c("auto","hetero","detritus")},
indicating whether each taxon is a primary producer, consumer/predator, or
detritus, respectively.}

\item{U}{numeric-vector with names matching \code{taxa}, providing the
proportion of consumption that is unassimilated and therefore
exported to detritus}

\item{fit_B}{Character-vector listing \code{taxa} for which equilibrium
biomass is estimated as a fixed effect}

\item{fit_Q}{Character-vector listing \code{taxa} for which the catchability
coefficient is estimated as a fixed effect}

\item{fit_B0}{Character-vector listing \code{taxa} for which the ratio of initial
to equilibrium biomass is estimated as a fixed effect}

\item{fit_eps}{Character-vector listing \code{taxa} for which the
model should estimate annual process errors in dB/dt}

\item{control}{Output from \code{\link[=ecostate_control]{ecostate_control()}}, used to define user
settings.}
}
\description{
Estimate parameters for an EcoState model
}
\details{
All \code{taxa} must be included in \code{QB}, \code{PB}, \code{B}, and \code{DC},
but additional taxa can be in \code{QB}, \code{PB}, \code{B}, and \code{DC} that
are not in \code{taxa}.  So \code{taxa} can be used to redefine the set of modeled
species without changing other inputs
}
