% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{elementwise_product}
\alias{elementwise_product}
\title{Elementwise product of sparse and dense matrices}
\usage{
elementwise_product(Msparse, Mdense)
}
\arguments{
\item{Msparse}{sparse matrix}

\item{Mdense}{dense matrix}
}
\description{
Calculate elementwise product of sparse and dense matrices
}
