% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{evaluate_prior}
\alias{evaluate_prior}
\title{EcoState priors}
\usage{
evaluate_prior(priors, p, taxa, years, stanza_groups, sem = "")
}
\arguments{
\item{priors}{Either:
\itemize{
\item A list of sampling statements (e.g., \code{logtau_i[] ~ dnorm(mean = log(0.5), sd = 1)}) for named parameters.
See "Details" for a list of parameters and their descriptions.
\item A function that takes as input the list of
parameters \code{out$obj$env$parList()} where \code{out} is the output from
\code{ecostate()}, and returns a numeric vector
where the sum is the log-prior probability.  For example
\code{log_prior = function(p) dnorm( p$logq_i[1], mean=0, sd=0.1, log=TRUE)}
specifies a lognormal prior probability for the catchability coefficient
for the first \code{taxa} with logmean of zero and logsd of 0.1
}}

\item{p}{List of parameters}

\item{taxa}{Character vector of taxa included in mode}

\item{years}{Integer-vector of years included in model}

\item{stanza_groups}{Character vector of unique multi-stanza groups}

\item{sem}{SEM model data frame}
}
\value{
Log-prior density (numeric)
}
\description{
EcoState priors
}
\details{
Priors can be specified as a list of sampling statements, with parameter names on the left-hand side and density
functions and their named arguments on the right-hand side. SEM parameters should be referred to by their names,
while all other parameters are indexed (e.g., by taxa, year, or stanza group) and should be specified with brackets.
E.g., the following list specifies priors on an SEM parameter named \code{rho_X}, the biomass process errors for all time
steps for taxa Y \code{epsilon_ti[,"Y"]}, and the ecotrophic efficiency for all taxa:
\preformatted{
 priors <- list(
   rho_X ~ dnorm(mean = 0, sd = 0.5), 
   epsilon_ti[,"Y"] ~ dnorm(mean = 0, sd = 0.1), 
   EE_i[] ~ dnorm(mean = 1, sd = 0.1)
 )
}
}
