% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_trophic_level}
\alias{get_trophic_level}
\title{Calculate trophic level}
\usage{
get_trophic_level(Q_ij)
}
\arguments{
\item{Q_ij}{Consumption of each prey i by predator j.}
}
\description{
Calculate trophic level from consumption matrix
}
