% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_mle}
\alias{goa_mle}
\title{Output from optimizer for GOA stanza example}
\usage{
data(goa_mle)
}
\description{
Objective function and MLE for coefficients in GOA stanza example
}
\keyword{data}
