% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myode.R
\name{myode}
\alias{myode}
\title{Generic ODE solver}
\usage{
myode(f, a, b, y0, n, Pars, method)
}
\arguments{
\item{f}{function in the differential equation y' = f(x, y);
defined as a function R \times R^m \rightarrow R^m, where m is the number of equations.}

\item{a}{starting time for the interval to integrate}

\item{b}{ending time for the interval to integrate.}

\item{y0}{starting values at time \code{a}}

\item{Pars}{named list of parameters passed to f}

\item{steps}{the number of steps from a to b.}
}
\description{
Interface for \code{RTMBode::ode}, itself an interface for \code{deSolve::ode}
}
\details{
Todo
}
