% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_agecomp}
\alias{plot_agecomp}
\title{Plot fits to age composition data}
\usage{
plot_agecomp(model, stgroup, years, track_cohorts = TRUE, cutoff = 0)
}
\arguments{
\item{model}{Fitted model returned by \code{ecostate}}

\item{stgroup}{Stanza group to plot age composition fits for. Must be length 1.
Defaults to the first stanza group in \code{settings$unique_stanza_groups}.}

\item{years}{Years to plot. Defaults to all years with age composition data}

\item{track_cohorts}{Whether to color observed agecomp data by cohort. Defaults to TRUE}

\item{cutoff}{Minimum proportion for plotting an age bin. Specifying a cutoff may
reduce the number of cohorts and make cohort colors easier to discern.}
}
\value{
A ggplot object
}
\description{
Plot fits to age composition data
}
