% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostate.R
\name{print.ecostate}
\alias{print.ecostate}
\title{Print fitted ecostate object}
\usage{
\method{print}{ecostate}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{ecostate}}}

\item{...}{Not used}
}
\value{
No return value, called to provide clean terminal output when calling fitted
object in terminal.
}
\description{
Prints output from fitted ecostate model
}
