% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostate.R
\name{print_ecopars}
\alias{print_ecopars}
\title{Print EcoSim parameters}
\usage{
print_ecopars(x, silent = FALSE)
}
\arguments{
\item{x}{Output from \code{\link{ecostate}}}

\item{silent}{whether to print to terminal}
}
\value{
invisibly returns table printed
}
\description{
Prints parameters defining EcoSim dynamics
}
