% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4sys.R
\name{rk4sys}
\alias{rk4sys}
\title{Classical Runge-Kutta for system of equations}
\usage{
rk4sys(f, a, b, y0, n, Pars)
}
\arguments{
\item{f}{function in the differential equation y' = f(x, y);
defined as a function R \times R^m \rightarrow R^m, where m is the number of equations.}

\item{a}{starting time for the interval to integrate}

\item{b}{ending time for the interval to integrate.}

\item{y0}{starting values at time \code{a}}

\item{Pars}{named list of parameters passed to f}

\item{steps}{the number of steps from a to b.}
}
\description{
Classical Runge-Kutta of order 4.
}
\details{
Classical Runge-Kutta of order 4 for (systems of) ordinary differential equations with fixed step size.
Copied from pracma under GPL-3, with small modifications to work with RTMB
}
