% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanzas.R
\name{stanza_settings}
\alias{stanza_settings}
\title{Detailed control for stanza structure}
\usage{
stanza_settings(
  taxa,
  stanza_groups,
  K,
  d,
  Wmat,
  Amax,
  SpawnX,
  Leading,
  fit_K = c(),
  fit_d = c(),
  Amat = NULL,
  Wmatslope,
  STEPS_PER_YEAR = 1,
  comp_weight = c("multinom", "dir", "dirmult"),
  correct_errors = FALSE,
  min_agecomp_prob = 0
)
}
\description{
Define a list of control parameters.
}
