% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whitehouse_2021}
\alias{whitehouse_2021}
\title{Full rpath inputs for eastern Bering Sea}
\usage{
data(whitehouse_2021)
}
\description{
All Rpath inputs from Whitehouse et al. 2021
}
\keyword{data}
