% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_equilibrium.R
\name{add_equilibrium}
\alias{add_equilibrium}
\title{Compute equilibrium values}
\usage{
add_equilibrium(ecoparams)
}
\arguments{
\item{p}{list of parameters}
}
\description{
Compute equilibrium values
}
\details{
todo
}
