% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecostate.R
\name{ecostate_control}
\alias{ecostate_control}
\title{Detailed control for ecostate structure}
\usage{
ecostate_control(
  nlminb_loops = 1,
  newton_loops = 0,
  eval.max = 1000,
  iter.max = 1000,
  getsd = TRUE,
  silent = getOption("ecostate.silent", TRUE),
  trace = getOption("ecostate.trace", 0),
  verbose = getOption("ecostate.verbose", FALSE),
  profile = c(),
  tmb_par = NULL,
  getJointPrecision = FALSE
)
}
\arguments{
\item{nlminb_loops}{Integer number of times to call [stats::nlminb()].}

\item{newton_loops}{Integer number of Newton steps to do after running
[stats::nlminb()].}

\item{eval.max}{Maximum number of evaluations of the objective function
allowed. Passed to `control` in [stats::nlminb()].}

\item{iter.max}{Maximum number of iterations allowed. Passed to `control` in
[stats::nlminb()].}

\item{getsd}{Boolean indicating whether to call [TMB::sdreport()]}

\item{silent}{Disable terminal output for inner optimizer?}

\item{trace}{Parameter values are printed every `trace` iteration
for the outer optimizer. Passed to
`control` in [stats::nlminb()].}

\item{verbose}{Output additional messages about model steps during fitting?}

\item{tmb_par}{list of parameters for starting values, with shape identical
to `tinyVAST(...)$internal$parlist`}

\item{getJointPrecision}{whether to get the joint precision matrix.  Passed
to \code{\link[TMB]{sdreport}}.}
}
\description{
Define a list of control parameters.
}
