% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_trophic_level}
\alias{get_trophic_level}
\title{Calculate trophic level}
\usage{
get_trophic_level(
  Q_ij,
  inverse_method = c("Penrose_moore", "Standard"),
  which_primary = which(colSums(Q_ij) == 0)
)
}
\arguments{
\item{Q_ij}{Consumption of each prey i by predator j in units biomass.}

\item{inverse_method}{whether to use pseudoinverse or standard inverse}
}
\description{
Calculate trophic level from consumption matrix
}
\details{
Trophic level \eqn{l_i} for each predator \eqn{i} is defined as:

\deqn{ \mathbf{l - 1 = l Q^*} }

where \eqn{\mathbf{Q*}} is the proportion consumption for each predator (column)
of different prey (rows).  We identify primary producers as any taxa with no
consumption (a column of 0s), and assign them as the first trophic level.
}
