% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot biomass time series from fitted ecostate model}
\usage{
plot_timeseries(fit, taxa, observed = TRUE, interval = 0.95, q_adj = TRUE)
}
\arguments{
\item{fit}{Fitted model object returned by \code{ecostate}}

\item{taxa}{Taxon names to plot biomass for. If missing, defaults to all modeled taxa.}

\item{observed}{Plot survey biomass for available taxa? Defaults to TRUE.}

\item{interval}{Coverage of plotted confidence bands. Defaults to 0.95 (95\%), omitted if FALSE.}

\item{q_adj}{Where applicable, should estimates adjusted for catchability be plotted? Useful for
assessing fit to survey data.}
}
\value{
invisibly return \code{ggplot} object for biomass time series
}
\description{
Plot biomass time series from fitted ecostate model
}
