% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project ecostate model to future times (EXPERIMENTAL)}
\usage{
project(
  model,
  nsim = 1,
  extra_years,
  Frate,
  covariates,
  future_var = TRUE,
  past_var = FALSE,
  parm_var = FALSE,
  obj_new
)
}
\arguments{
\item{model}{fitted model from \code{ecostate}}

\item{nsim}{Number simulations}

\item{extra_years}{a vector of extra years, matching values in \code{Frate} and
\code{covariates}}

\item{Frate}{matrix with rownames corresponding to \code{extra_years} and colnames
corresponding to \code{model$internal$taxa} giving fishing mortality in future
years. Any missing years or taxa will be assigned an F rate of 0.}

\item{covariates}{matrix with rownames corresponding to \code{extra_years}
and columns corresponding to SEM covariates. Covariates corresponding to
missing covariate columns and/or NA values within included covariate columns
will be estimated conditional on included covariates.}

\item{future_var}{logical indicating whether to simulate future process errors
from GMRF, or just compute the predictive mean}

\item{past_var}{logical indicating whether to re-simulate past process errors
from predictive distribution of random effects, thus changing the boundary
condition of the forecast}

\item{parm_var}{logical indicating whether to re-sample fixed effects from their
predictive distribution, thus changing the GMRF for future process errors}

\item{obj_new}{Optional; a re-built model object with parameter blocks covering
the full range of years in the simulation. Intended for internal use.}
}
\value{
A list (or list of lists, if nsim > 1) of estimated parameters and
derived quantities with structure resembling \code{model$internal$parhat}
}
\description{
Projects a fitted model forward in time, optionally
conditioning on covariates and/or catches.
}
